requirements.txt
python-telegram-bot==20.1
apscheduler==3.8.1
requests==2.28.1
pytz==2023.2
